requirements.txt
discord.py>=2.4.0
python-dotenv==1.0.0
aiohttp>=3.8.0
duckduckgo-search
google-generativeai
flask
